Refactor `revalidateWallets` to be declared once using context
